\name{Blendstat}

\alias{Blendstat}

\docType{package}

\title{Package for joint analysis of experiments with mixtures and random effects.}

\description{Package to perform a joint analysis of experiments with mixtures and random effects, assuming a process variable, represented by a covariate.}

\details{
\tabular{ll}{
Package: \tab Blendstat\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2018-02-20\cr
License: \tab GPL(>= 2)\cr
LazyLoad: \tab yes\cr
}
}

\author{
Marcelo Angelo Cirillo e Paulo Cesar Ossani.

Maintainer: Paulo Cesar Ossani <ossanipc@hotmail.com>
}

\references{
Kalirajan, K. P.. On the estimation of a regression model with fixed and random coefficients. \emph{Journal of Applied Statistics}, 17: 237-244, 1990. doi:10.1080/757582835
}

\keyword{Optimization}
