\name{Blendstat}

\alias{Blendstat}

\docType{package}

\title{A R Package for Joint Analysis of Experiments with Mixtures and Random Effects.}

\description{Joint analysis of experiments with mixtures and random effects, taking on a process variable represented by a covariable.}

\details{
\tabular{ll}{
Package: \tab Blendstat\cr
Type: \tab Package\cr
Version: \tab 1.0.2\cr
Date: \tab 2018-07-02\cr
License: \tab GPL(>= 2)\cr
LazyLoad: \tab yes\cr
}
}

\author{
Marcelo Angelo Cirillo and Paulo Cesar Ossani.

Maintainer: Paulo Cesar Ossani <ossanipc@hotmail.com>
}

\references{
Kalirajan, K. P. On the estimation of a regression model with fixed and random coefficients. \emph{Journal of Applied Statistics}, 17(2): 237-244, 1990. doi:10.1080/757582835

Swany, P. A. V. B. \emph{Statistical Inference in Random Coefficient Regression Models.} Amsterdam: Springer Science & Business Media, 1971. 209 p.
}

\keyword{Optimization}
